LayoutTests/fast/dom/Element/scrolling-funtions-on-body-expected.txt

Tests scrollTo/By(x,y) and scrollTo/By(scrollOptions) on the body in standard mode.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS scroller.scrollLeft is 0
PASS scroller.scrollTop is 0
PASS window.scrollX is 0
PASS window.scrollY is 0
PASS scroller.scrollLeft is 0
PASS scroller.scrollTop is 0
PASS window.scrollX is 0
PASS window.scrollY is 0
PASS scroller.scrollLeft is 0
PASS scroller.scrollTop is 0
PASS window.scrollX is 0
PASS window.scrollY is 0
PASS successfullyParsed is true

TEST COMPLETE
